Remove `$effect` altogether, since we don't need it.
Currently, if `boardApiInfo` would be changed for example outside of Board.svelte after the Board component is rendered, the Board component will re-render, but the `<script>` of the Board component will not re-run => the issue would be that the `naTahu` will not be computed again.

The way to fix this would be to track uuid changes and bind an effect to that.

In the current usage, we don't care, as we destroy the component after every game.